Implemented login view for logging of the user
Implemented login view with components like email/user text field, password field, background image, so on. Added a small check if the user/pass is admin/admin and if so, the user gets forwarded to a new layout.  If the credentials are wrong, the user receives a message with "wrong credentials".